+++ src/block/array.ts
@@ -2,7 +2,7 @@ import * as Blockly from 'blockly'
 // @ts-ignore
 import { BlockDefinition } from 'blockly/core/blocks'
 import { calciumVariableName } from './variable'
-import { allTypesForCheck } from '../type-check/all-types'
+import { allTypesForCheck } from '../block/type-check/all-types'
 
 export const pseudoAssignArrayName = 'pseudo_assign_array'
 export const pseudoAssignArrayMutatorName = 'pseudo_assign_array_mutator'
