+++ src/block/variable.ts
