+++ src/lang/ja-jp/message.ts
