+++ src/lang/ja-jp/tooltip.ts
